# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Call Attribution Platform API",
    "description": "Deterministic call attribution: number allocation, ingestion, qualification, publication and administration. See specs/001-call-attribution-platform/contracts/ for the narrative wire contracts this schema implements.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/admin/pools": {
      "post": {
        "tags": [
          "AdminPools"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoolRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoolRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoolRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "AdminPools"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/pools/{id}": {
      "get": {
        "tags": [
          "AdminPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/pools/{id}/numbers": {
      "get": {
        "tags": [
          "AdminPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/pools/{id}/numbers/import": {
      "post": {
        "tags": [
          "AdminPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/admin/pools/{id}/numbers/import-from-folder": {
      "post": {
        "tags": [
          "AdminPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportFromFolderRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportFromFolderRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportFromFolderRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/admin/qualification-rules": {
      "get": {
        "tags": [
          "AdminQualificationRules"
        ],
        "parameters": [
          {
            "name": "scope_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope_ref",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "AdminQualificationRules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQualificationRuleRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQualificationRuleRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQualificationRuleRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/admin/qualification-rules/{id}": {
      "delete": {
        "tags": [
          "AdminQualificationRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePoolRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "scope_type": {
            "type": "string",
            "nullable": true
          },
          "scope_ref": {
            "type": "string",
            "nullable": true
          },
          "default_number": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateQualificationRuleRequestDto": {
        "type": "object",
        "properties": {
          "scope_type": {
            "type": "string",
            "nullable": true
          },
          "scope_ref": {
            "type": "string",
            "nullable": true
          },
          "conditions": {
            "$ref": "#/components/schemas/QualificationConditionsDto"
          },
          "effective_start": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ImportFromFolderRequestDto": {
        "type": "object",
        "properties": {
          "file_name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QualificationConditionsDto": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "nullable": true
          },
          "answered_required": {
            "type": "boolean"
          },
          "min_connected_duration_seconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "time_of_day": {
            "$ref": "#/components/schemas/TimeOfDayDto"
          }
        },
        "additionalProperties": false
      },
...
      "TimeOfDayDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "nullable": true
          },
          "end": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}